% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{darkTriad}
\alias{darkTriad}
\title{Short Dark Triad}
\format{
## `darkTriad`
A data frame with 18192 rows and 29 columns:
\describe{
  \item{M}{Machiavellianism items}
  \item{N}{Narcissism items}
  \item{P}{Psychopathy}
  \item{Country}{Participant's Country}
  \item{Source}{Source}

}
}
\source{
<https://openpsychometrics.org/_rawdata/>
}
\usage{
darkTriad
}
\description{
Data set taken from the Open-Source Psychometrics Project.
It is a data frame with 27 columns representing
the 27 items of the instrument plus two extra columns.
}
\keyword{datasets}
