% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataRum}
\alias{dataRum}
\title{Rumination study}
\format{
## `dataRum`
A data frame with 212 rows and 45 columns:
\describe{
  \item{pending}{pending}


}
}
\usage{
dataRum
}
\description{
A study conducted by David Campos-Miranda & Esteban Montenegro-Montenegro
in Costa Rica. This data set is a fraction of the data comprising
information about depression symptoms, metacognitive beliefs, rumination,
worry, and negative priming experiments. The study was conducted in 2011.
Participants were teenagers living in the Central Valley.
}
\keyword{datasets}
