% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ouesData}
\alias{ouesData}
\title{EDAD study Costa Rica and Kansas}
\format{
## `ouesData`
A data frame with 212 rows and 45 columns:
\describe{
  \item{pending}{pending}


}
}
\usage{
ouesData
}
\description{
A study conducted by David J. Johnson, Monica Salazar-Villanea,
Jose Moncada, and other collaborators from Universuty of Kansas,
University of California, Davis, and University of Costa Rica.
}
\details{
The main outcome in this study was the Oxygen uptake efficiency slope (OUES),
in a sample of aging participants from Costa Rica and Kansas.
}
\keyword{datasets}
